Move proto files out of 2d/3d directories.
This is to make it possible to use the proto definitions
from Python where the directory structure otherwise
prevents them from being imported in the usual way.
