+++ cartographer/mapping/2d/grid_2d.h
@@ -22,9 +22,9 @@
 #include "cartographer/mapping/2d/map_limits.h"
 #include "cartographer/mapping/grid_interface.h"
 #include "cartographer/mapping/probability_values.h"
-#include "cartographer/mapping/proto/2d/grid_2d.pb.h"
-#include "cartographer/mapping/proto/2d/submaps_options_2d.pb.h"
+#include "cartographer/mapping/proto/grid_2d.pb.h"
 #include "cartographer/mapping/proto/submap_visualization.pb.h"
+#include "cartographer/mapping/proto/submaps_options_2d.pb.h"
 #include "cartographer/mapping/value_conversion_tables.h"
 
 namespace cartographer {
