+++ cartographer/mapping/internal/2d/local_trajectory_builder_2d.h
@@ -27,7 +27,7 @@
 #include "cartographer/mapping/internal/motion_filter.h"
 #include "cartographer/mapping/internal/range_data_collator.h"
 #include "cartographer/mapping/pose_extrapolator.h"
-#include "cartographer/mapping/proto/2d/local_trajectory_builder_options_2d.pb.h"
+#include "cartographer/mapping/proto/local_trajectory_builder_options_2d.pb.h"
 #include "cartographer/metrics/family_factory.h"
 #include "cartographer/sensor/imu_data.h"
 #include "cartographer/sensor/internal/voxel_filter.h"
