+++ cartographer/mapping/internal/2d/local_trajectory_builder_options_2d.h
@@ -18,7 +18,7 @@
 #define CARTOGRAPHER_MAPPING_INTERNAL_2D_LOCAL_TRAJECTORY_BUILDER_OPTIONS_2D_H_
 
 #include "cartographer/common/lua_parameter_dictionary.h"
-#include "cartographer/mapping/proto/2d/local_trajectory_builder_options_2d.pb.h"
+#include "cartographer/mapping/proto/local_trajectory_builder_options_2d.pb.h"
 
 namespace cartographer {
 namespace mapping {
