+++ cartographer/mapping/internal/2d/normal_estimation_2d.h
@@ -19,7 +19,7 @@
 
 #include <vector>
 
-#include "cartographer/mapping/proto/2d/normal_estimation_options_2d.pb.h"
+#include "cartographer/mapping/proto/normal_estimation_options_2d.pb.h"
 #include "cartographer/sensor/point_cloud.h"
 #include "cartographer/sensor/range_data.h"
 #include "cartographer/transform/transform.h"
