+++ cartographer/mapping/internal/3d/local_trajectory_builder_3d.cc
@@ -21,10 +21,10 @@
 #include "absl/memory/memory.h"
 #include "cartographer/common/time.h"
 #include "cartographer/mapping/internal/3d/scan_matching/rotational_scan_matcher.h"
-#include "cartographer/mapping/proto/3d/local_trajectory_builder_options_3d.pb.h"
-#include "cartographer/mapping/proto/3d/submaps_options_3d.pb.h"
-#include "cartographer/mapping/proto/scan_matching//ceres_scan_matcher_options_3d.pb.h"
-#include "cartographer/mapping/proto/scan_matching//real_time_correlative_scan_matcher_options.pb.h"
+#include "cartographer/mapping/proto/local_trajectory_builder_options_3d.pb.h"
+#include "cartographer/mapping/proto/scan_matching/ceres_scan_matcher_options_3d.pb.h"
+#include "cartographer/mapping/proto/scan_matching/real_time_correlative_scan_matcher_options.pb.h"
+#include "cartographer/mapping/proto/submaps_options_3d.pb.h"
 #include "cartographer/transform/timestamped_transform.h"
 #include "glog/logging.h"
