+++ cartographer/mapping/internal/3d/local_trajectory_builder_3d.h
@@ -27,7 +27,7 @@
 #include "cartographer/mapping/internal/motion_filter.h"
 #include "cartographer/mapping/internal/range_data_collator.h"
 #include "cartographer/mapping/pose_extrapolator_interface.h"
-#include "cartographer/mapping/proto/3d/local_trajectory_builder_options_3d.pb.h"
+#include "cartographer/mapping/proto/local_trajectory_builder_options_3d.pb.h"
 #include "cartographer/metrics/family_factory.h"
 #include "cartographer/sensor/imu_data.h"
 #include "cartographer/sensor/internal/voxel_filter.h"
