+++ cartographer/mapping/proto/grid_2d.proto
@@ -16,9 +16,9 @@ syntax = "proto3";
 
 package cartographer.mapping.proto;
 
-import "cartographer/mapping/proto/2d/map_limits.proto";
-import "cartographer/mapping/proto/2d/probability_grid.proto";
-import "cartographer/mapping/proto/2d/tsdf_2d.proto";
+import "cartographer/mapping/proto/map_limits.proto";
+import "cartographer/mapping/proto/probability_grid.proto";
+import "cartographer/mapping/proto/tsdf_2d.proto";
 
 message Grid2D {
   message CellBox {
