+++ cartographer/mapping/proto/grid_2d_options.proto
