+++ cartographer/mapping/proto/local_trajectory_builder_options_2d.proto
@@ -19,9 +19,9 @@ package cartographer.mapping.proto;
 import "cartographer/mapping/proto/motion_filter_options.proto";
 import "cartographer/mapping/proto/pose_extrapolator_options.proto";
 import "cartographer/sensor/proto/adaptive_voxel_filter_options.proto";
-import "cartographer/mapping/proto/2d/submaps_options_2d.proto";
 import "cartographer/mapping/proto/scan_matching/ceres_scan_matcher_options_2d.proto";
 import "cartographer/mapping/proto/scan_matching/real_time_correlative_scan_matcher_options.proto";
+import "cartographer/mapping/proto/submaps_options_2d.proto";
 
 // NEXT ID: 22
 message LocalTrajectoryBuilderOptions2D {
