+++ cartographer/mapping/proto/local_trajectory_builder_options_3d.proto
@@ -16,11 +16,11 @@ syntax = "proto3";
 
 package cartographer.mapping.proto;
 
-import "cartographer/mapping/proto/3d/submaps_options_3d.proto";
 import "cartographer/mapping/proto/motion_filter_options.proto";
 import "cartographer/mapping/proto/pose_extrapolator_options.proto";
 import "cartographer/mapping/proto/scan_matching/ceres_scan_matcher_options_3d.proto";
 import "cartographer/mapping/proto/scan_matching/real_time_correlative_scan_matcher_options.proto";
+import "cartographer/mapping/proto/submaps_options_3d.proto";
 import "cartographer/sensor/proto/adaptive_voxel_filter_options.proto";
 import "cartographer/sensor/proto/sensor.proto";
 import "cartographer/transform/proto/timestamped_transform.proto";
