+++ cartographer/mapping/proto/submap.proto
@@ -16,8 +16,8 @@ syntax = "proto3";
 
 package cartographer.mapping.proto;
 
-import "cartographer/mapping/proto/2d/grid_2d.proto";
-import "cartographer/mapping/proto/3d/hybrid_grid.proto";
+import "cartographer/mapping/proto/grid_2d.proto";
+import "cartographer/mapping/proto/hybrid_grid.proto";
 import "cartographer/transform/proto/transform.proto";
 
 // Serialized state of a Submap2D.
