+++ cartographer/mapping/proto/submaps_options_3d.proto
@@ -14,16 +14,27 @@
 
 syntax = "proto3";
 
-import "cartographer/mapping/proto/2d/grid_2d_options.proto";
-import "cartographer/mapping/proto/range_data_inserter_options.proto";
+import "cartographer/mapping/proto/range_data_inserter_options_3d.proto";
 
 package cartographer.mapping.proto;
 
-message SubmapsOptions2D {
+message SubmapsOptions3D {
+  // Resolution of the 'high_resolution' map in meters used for local SLAM and
+  // loop closure.
+  double high_resolution = 1;
+
+  // Maximum range to filter the point cloud to before insertion into the
+  // 'high_resolution' map.
+  double high_resolution_max_range = 4;
+
+  // Resolution of the 'low_resolution' version of the map in meters used for
+  // local SLAM only.
+  double low_resolution = 5;
+
   // Number of range data before adding a new submap. Each submap will get twice
   // the number of range data inserted: First for initialization without being
   // matched against, then while being matched.
-  int32 num_range_data = 1;
-  GridOptions2D grid_options_2d = 2;
-  RangeDataInserterOptions range_data_inserter_options = 3;
+  int32 num_range_data = 2;
+
+  RangeDataInserterOptions3D range_data_inserter_options = 3;
 }
